Abstract data_storage
Abstract data_storage
